backend/requirements.txt
langchain == 0.1.19
langchain_google_genai
langchain-chroma
fastapi == 0.111.1
uvicorn == 0.30.1
python-dotenv
pydantic == 2.8.2
chromadb
langsmith
scikit-learn
pandas
plotly